Validators are implemented using Fluent Validation

+++ Deal Management System/DTOs/HotelDTO.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
 
-        public double Rate { get; set; }
+        public decimal Rate { get; set; }
         public string Amenities { get; set; }
     }
 }
